requirements.txt
requests==2.31.0
langchain==0.0.353
streamlit==1.32.0
tiktoken==0.5.2
qdrant-client==1.7.0
streamlit-chat==0.1.1
openai==0.28.1
python-dotenv==0.19.0
